prj: add simple classification using only CLS

+++ project/utils/data.py
@@ -255,9 +255,7 @@ class StockDataset(torch.utils.data.Dataset):
 
     def __getitem__(self, idx):
         item = {key: torch.tensor(val[idx]) for key, val in self.encodings.items()}
-        item['labels'] = torch.tensor(self.labels[idx], dtype=torch.float)
-        # if item['labels'].dtype != torch.float16:
-        #     print("error")
+        item['labels'] = torch.tensor(self.labels[idx])
         return item
 
     def __len__(self):
